Feat: add bash and zsh completion files
This adds completions file for bash and zsh.
The setup.py is changed so the files are being installed to
- <root>/share/bash-completion/completions/cppman.bash
- <root>/share/zsh-completion/completions/cppman.zsh

Bash treats ':' as white spaces which makes using bash completion not as
nice as I wished (zsh does not have this issue). This can be fixed for
bash by adding the following line to `~/.bashrc`

```
export COMP_WORDBREAKS=" /\"\'><;|&("
```