Fix: crawler respecting max_depth correctly
The crawler wasn't using max_depth correctly. It was using the directory
depth to determine the depth which is flawed.

Now the crawler keeps track of the depth of each link and page. It is
important to do BFS search depending on the depth, otherwise the depth
copmutation might be wrong.