Fix: all searches should be case sensitive
All important names in c++ for keywords and classes are lowercase. There
are a few entries in the database like 'ATOMIC_xxx_LOCK_FREE' that get
called when running `cppman atomic` which is not as intended.
To fix this behavior I added case sensitivity when running searches on
the database.