Fix look up for std::string::string
it wasn't possible to lookup `std::string::string` (Constructor of
string)
The issue is that `std::string` is an alias of `std::basic_string`.
The way we resolved aliases was to add extra keywords.
This would only produce keywords to lookup `std::string::basic_string`
but not `std::string::string`.

This commit adds a `_commonprefix` function to find the prefix of two
strings.
If e.g. we try to replace `std::basic_string` with `std::string` we check if
there is a common prefix that ends with "::" and will also do a
replacement of "basic_string" to "string".