fix: remove [static] marker from std::string::npos
When viewing std::string it wasn't possible to follow the link
to `std::string::npos`. The issue is that it was marked with
`npos[static] (3)` which is not a valid link.

Also it was missing the prefix of the class.

This commit fixes these issue. `std::string` will now show a proper link
to `std::string::npos`